Made keyNumber a byte to save space

+++ Keypad_Decoder.ino
@@ -54,7 +54,7 @@ void setup ()
 
 void loop ()
   {
-  int keyNumber = 0;
+  byte keyNumber = 0;
   unsigned long now = millis ();  // for debouncing
   
   // check each row
